Add `/mnt/physicsdemos-data` to docker-compose.yml
The physicsdemos data directory is expected to be in the address
`/mnt/physicsdemos-data`, before this commit it just wasn't there.

Now docker expects to find this directory in the root of the project,
and it automatically mounts it inside the container in the correct
address.

If the directory doesn't exists docker creates it empty, it causes
no problem on the website but the videos and thumbnails will not
show up.