Improve setup process to work in a single command
Now it is possible to setup the docker project using a single
recipe from the Makefile:

- `make setup`

It should setup the website, the docker images and the database.

Then you may run `make up` and the website should be
available at the address `http://localhost:80`.

After this command the website will be working but you will still
need to download or copy the physicsdemos-data directory and
save it in the root of the project so docker can mount it inside
the container otherwise the videos and thumbnails will not work.