Add shortcuts for running cake and composer
To run a command inside the physicsdemos.app container it is
required a sligtly long command:

docker exec -it physicsdemos.app ...

These 2 scripts encapsulate the commands so you can use it like:

```
./cake migrations migrate
./cake composer install
```

The composer script is not working so well because it complains
about being executed as root user.

I will have to search how to fix this in a future commit.